requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
Django==3.2.16
django-dynamic-admin-forms==0.1.8
django-kronos==1.0
django-model-utils==4.2.0
django-notifications-hq==1.7.0
idna==3.4
jsonfield==3.1.0
Pillow==9.2.0
psycopg2==2.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.5
qrcode==7.3.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
swapper==1.3.0
tzdata==2022.4
urllib3==1.26.12
